refacor: refactor arch
