+++ internal/core/port/service/user.go
@@ -6,8 +6,6 @@ import (
 	"fmt"
 
 	"github.com/ezex-io/ezex-users/internal/core/entity"
-	"github.com/ezex-io/ezex-users/internal/core/model/request"
-	"github.com/ezex-io/ezex-users/internal/core/model/response"
 	"github.com/ezex-io/ezex-users/internal/core/port/repository"
 )
 
@@ -16,14 +14,14 @@ type UserService interface {
 	// SaveSecurityImage saves a security image for a user.
 	SaveSecurityImage(
 		ctx context.Context,
-		req *request.SaveSecurityImageRequest,
-	) (*response.SaveSecurityImageResponse, error)
+		req *SaveSecurityImageRequest,
+	) (*SaveSecurityImageResponse, error)
 
 	// GetSecurityImage retrieves a security image by UserID.
 	GetSecurityImage(
 		ctx context.Context,
-		req *request.GetSecurityImageRequest,
-	) (*response.GetSecurityImageResponse, error)
+		req *GetSecurityImageRequest,
+	) (*GetSecurityImageResponse, error)
 }
 
 func NewUserService(repo repository.Repository) UserService {
@@ -36,8 +34,8 @@ type userService struct {
 
 func (s *userService) SaveSecurityImage(
 	ctx context.Context,
-	req *request.SaveSecurityImageRequest,
-) (*response.SaveSecurityImageResponse, error) {
+	req *SaveSecurityImageRequest,
+) (*SaveSecurityImageResponse, error) {
 	err := s.repo.SecurityImage().Save(ctx, &entity.SecurityImage{
 		UserID:         req.UserID,
 		SecurityImage:  req.SecurityImage,
@@ -47,19 +45,19 @@ func (s *userService) SaveSecurityImage(
 		return nil, fmt.Errorf("failed to save security image: %w", err)
 	}
 
-	return &response.SaveSecurityImageResponse{}, nil
+	return &SaveSecurityImageResponse{}, nil
 }
 
 func (s *userService) GetSecurityImage(
 	ctx context.Context,
-	req *request.GetSecurityImageRequest,
-) (*response.GetSecurityImageResponse, error) {
+	req *GetSecurityImageRequest,
+) (*GetSecurityImageResponse, error) {
 	image, err := s.repo.SecurityImage().GetByID(ctx, req.UserID)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get security image: %w", err)
 	}
 
-	return &response.GetSecurityImageResponse{
+	return &GetSecurityImageResponse{
 		UserID:         image.UserID,
 		SecurityImage:  image.SecurityImage,
 		SecurityPhrase: image.SecurityPhrase,
